requirements-lock.txt
cffi==1.15.1
click==8.1.3
Flask==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.1
pycparser==2.21
pygit2==1.12.0
pypandoc==1.7.0
setuptools==64.0.3
Werkzeug==2.2.3
wheel==0.40.0
requests-cache
PyGithub==1.58.1
envparse==0.2.0
jira==3.5.0
